requirements.txt
Django==1.6
PIL==1.1.7
South==0.8.4
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
djangorestframework==2.3.9
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
